feat/ add relation on user to recipe and ingredient

+++ Backend_planifer_miam/src/Controller/IngredientController.php
@@ -9,22 +9,27 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
+#[IsGranted('IS_AUTHENTICATED_FULLY')]
 class IngredientController extends AbstractController
 {
     #[Route('/api/ingredients', methods: ['GET'])]
     public function getAll(Request $request, IngredientRepository $ingredientRepository): JsonResponse
     {
+        $user = $this->getUser();
         $search = $request->query->get('search');
+
+        $qb = $ingredientRepository->createQueryBuilder('i')
+            ->where('i.user = :user')
+            ->setParameter('user', $user);
+
         if ($search) {
-            $ingredients = $ingredientRepository->createQueryBuilder('i')
-                ->where('i.nameIngredient LIKE :search')
-                ->setParameter('search', '%' . $search . '%')
-                ->getQuery()
-                ->getResult();
-        } else {
-            $ingredients = $ingredientRepository->findAll();
+            $qb->andWhere('i.nameIngredient LIKE :search')
+               ->setParameter('search', '%' . $search . '%');
         }
+
+        $ingredients = $qb->getQuery()->getResult();
 
         return $this->json($ingredients, 200, [], ['groups' => 'ingredient:read']);
     }
@@ -32,6 +37,8 @@ class IngredientController extends AbstractController
     #[Route('/api/ingredients/{id}', methods: ['GET'])]
     public function getOne(Ingredient $ingredient): JsonResponse
     {
+        $this->denyAccessUnlessGranted('view', $ingredient);
+
         return $this->json($ingredient, 200, [], ['groups' => 'ingredient:read']);
     }
 
@@ -47,6 +54,7 @@ class IngredientController extends AbstractController
         $ingredient = new Ingredient();
         $ingredient->setNameIngredient($data['name_ingredient']);
         $ingredient->setUnit($data['unit']);
+        $ingredient->setUser($this->getUser());
 
         $em->persist($ingredient);
         $em->flush();
@@ -57,6 +65,8 @@ class IngredientController extends AbstractController
     #[Route('/api/ingredients/{id}', methods: ['PUT'])]
     public function update(Request $request, Ingredient $ingredient, EntityManagerInterface $em): JsonResponse
     {
+        $this->denyAccessUnlessGranted('edit', $ingredient);
+
         $data = json_decode($request->getContent(), true);
 
         if (isset($data['name_ingredient'])) {
@@ -75,6 +85,8 @@ class IngredientController extends AbstractController
     #[Route('/api/ingredients/{id}', methods: ['DELETE'])]
     public function delete(Ingredient $ingredient, EntityManagerInterface $em): JsonResponse
     {
+        $this->denyAccessUnlessGranted('delete', $ingredient);
+
         $em->remove($ingredient);
         $em->flush();
